1. передела все классы 2. пока работает авторизация и регистрация

+++ src/main/java/com/example/graduate_work_team2/repository/AdsRepository.java
@@ -13,6 +13,6 @@ import org.springframework.stereotype.Repository;
 
 /** Механизм для хранения, извлечения, обновления, удаления и поиска объектов. */
 @Repository
-public interface AdsRepository extends JpaRepository<Ads, Long> {
+public interface AdsRepository extends JpaRepository<Ads, Integer> {
 
 }
